+++ BE/src/app.module.ts
@@ -11,6 +11,7 @@ import { SocketService } from './websocket/socket.service';
 import { SocketGateway } from './websocket/socket.gateway';
 import { StockTopfiveModule } from './stock/topfive/stock.topfive.module';
 import { KoreaInvestmentModule } from './koreaInvestment/korea.investment.module';
+import { SocketModule } from './websocket/socket.module';
 
 @Module({
   imports: [
@@ -30,8 +31,9 @@ import { KoreaInvestmentModule } from './koreaInvestment/korea.investment.module
     AuthModule,
     StockIndexModule,
     StockTopfiveModule,
+    SocketModule,
   ],
   controllers: [AppController],
-  providers: [AppService, SocketService, SocketGateway],
+  providers: [AppService],
 })
 export class AppModule {}
