+++ BE/src/stock/index/interface/stock.index.interface.ts
@@ -1,10 +1,3 @@
-export interface AccessTokenInterface {
-  access_token: string;
-  access_token_token_expired: string;
-  token_type: string;
-  expires_in: number;
-}
-
 export interface StockIndexChartInterface {
   output: StockIndexChartElementInterface[];
   rt_cd: string;
