+++ BE/src/websocket/socket.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { SocketService } from './socket.service';
+import { SocketGateway } from './socket.gateway';
+import { KoreaInvestmentModule } from '../koreaInvestment/korea.investment.module';
+
+@Module({
+  imports: [KoreaInvestmentModule],
+  controllers: [],
+  providers: [SocketService, SocketGateway],
+  exports: [SocketGateway],
+})
+export class SocketModule {}
